bug fixed
The mismatch between the length of the  segment of  the original signal
and the size of the Hanning window(in CQT) leads to an unexpected
decrease in amplitude, especially for low frequency signals.

The problem is solved by duplicating the segment many times.

In addition, a few test .wav files are uploaded.